Add create/edit support
Add a new CAN_EDIT configuration parameter that can be used to enable
editing of quips and creation of new quips via the web interface.

This change also lays the groundwork for a voting feature by
introducing a scoring function based on number of up and down votes. The
scoring function chosen is from
http://www.evanmiller.org/how-not-to-sort-by-average-rating.html. This
will be used in a follow up commit to implement optional voting and
score ranked results.

Bug: #2
Bug: #3