help: Fix URL escaping of third example
The previous URL didn’t work for me, resulting in a 404 page; the new
URL is the result of typing the query into the search input and then
looking at document.location in the console. (The URL that Firefox shows
in the address bar doesn’t work when pasted into a new tab, since
Firefox partially decodes it in a way that Slim doesn’t seem to like.)

Tested in Firefox, Chromium, and GNOME Web (fka Epiphany), which ought
to cover the three major surviving browser engignes.